devdocs-backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# AI/ML - Embeddings
sentence-transformers==2.3.1
torch>=2.6.0
transformers==4.37.2

# Vector similarity (pgvector client)
pgvector==0.2.4

# CORS and security
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9

# Utilities
python-dotenv==1.0.0
httpx==0.26.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
